@@ -8,13 +8,19 @@
 'use strict';
 
 const withDefaults = require('../shared.webpack.config');
+const path = require('path');
 
 module.exports = withDefaults({
 	context: __dirname,
 	entry: {
-		extension: './src/ipynbMain.ts',
+		['ipynbMain.node']: './src/ipynbMain.node.ts',
+		notebookSerializerWorker: './src/notebookSerializerWorker.ts',
 	},
 	output: {
-		filename: 'ipynbMain.js'
-	}
+		path: path.resolve(__dirname, 'dist'),
+		filename: '[name].js'
+	},
+	plugins: [
+		...withDefaults.nodePlugins(__dirname), // add plugins, don't replace inherited
+	]
 });
