@@ -8,6 +8,7 @@ import { NotebookSerializer } from './notebookSerializer';
 import { activate as keepNotebookModelStoreInSync } from './notebookModelStoreSync';
 import { notebookImagePasteSetup } from './notebookImagePaste';
 import { AttachmentCleaner } from './notebookAttachmentCleaner';
+import { serializeNotebookToString } from './serializers';
 
 // From {nbformat.INotebookMetadata} in @jupyterlab/coreutils
 type NotebookMetadata = {
@@ -28,8 +29,8 @@ type NotebookMetadata = {
 	[propName: string]: unknown;
 };
 
-export function activate(context: vscode.ExtensionContext) {
-	const serializer = new NotebookSerializer(context);
+export function activate(context: vscode.ExtensionContext, isBrowser: boolean) {
+	const serializer = new NotebookSerializer(context, isBrowser);
 	keepNotebookModelStoreInSync(context);
 	context.subscriptions.push(vscode.workspace.registerNotebookSerializer('jupyter-notebook', serializer, {
 		transientOutputs: false,
@@ -105,8 +106,8 @@ export function activate(context: vscode.ExtensionContext) {
 		get dropCustomMetadata() {
 			return true;
 		},
-		exportNotebook: (notebook: vscode.NotebookData): string => {
-			return exportNotebook(notebook, serializer);
+		exportNotebook: (notebook: vscode.NotebookData): Promise<string> => {
+			return Promise.resolve(serializeNotebookToString(notebook));
 		},
 		setNotebookMetadata: async (resource: vscode.Uri, metadata: Partial<NotebookMetadata>): Promise<boolean> => {
 			const document = vscode.workspace.notebookDocuments.find(doc => doc.uri.toString() === resource.toString());
@@ -125,10 +126,6 @@ export function activate(context: vscode.ExtensionContext) {
 			return vscode.workspace.applyEdit(edit);
 		},
 	};
-}
-
-function exportNotebook(notebook: vscode.NotebookData, serializer: NotebookSerializer): string {
-	return serializer.serializeNotebookToString(notebook);
 }
 
 export function deactivate() { }
