@@ -506,8 +506,7 @@ class MyCompletionItem extends vscode.CompletionItem {
 		entry: Proto.CompletionEntry,
 		defaultCommitCharacters: readonly string[] | undefined,
 	): string[] | undefined {
-		// @ts-expect-error until TS 5.6
-		let commitCharacters = (entry.commitCharacters as string[] | undefined) ?? (defaultCommitCharacters ? Array.from(defaultCommitCharacters) : undefined);
+		let commitCharacters = entry.commitCharacters ?? (defaultCommitCharacters ? Array.from(defaultCommitCharacters) : undefined);
 		if (commitCharacters) {
 			if (context.enableCallCompletions
 				&& !context.isNewIdentifierLocation
@@ -781,8 +780,7 @@ class TypeScriptCompletionItemProvider implements vscode.CompletionItemProvider<
 		const isIncomplete = !!response.body.isIncomplete || (response.metadata as any)?.isIncomplete;
 		const entries = response.body.entries;
 		const metadata = response.metadata;
-		// @ts-expect-error until TS 5.6
-		const defaultCommitCharacters: readonly string[] | undefined = Object.freeze(response.body.defaultCommitCharacters);
+		const defaultCommitCharacters = Object.freeze(response.body.defaultCommitCharacters);
 
 		if (response.body.optionalReplacementSpan) {
 			optionalReplacementRange = typeConverters.Range.fromTextSpan(response.body.optionalReplacementSpan);
