@@ -4,15 +4,14 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as vscode from 'vscode';
+import { TypeScriptServiceConfiguration } from '../configuration/configuration';
 import { DiagnosticLanguage } from '../configuration/languageDescription';
+import { TelemetryReporter } from '../logging/telemetry';
+import { DiagnosticPerformanceData as TsDiagnosticPerformanceData } from '../tsServer/protocol/protocol';
 import * as arrays from '../utils/arrays';
 import { Disposable } from '../utils/dispose';
-import { ResourceMap } from '../utils/resourceMap';
-import { TelemetryReporter } from '../logging/telemetry';
-import { TypeScriptServiceConfiguration } from '../configuration/configuration';
 import { equals } from '../utils/objects';
-// @ts-expect-error until ts 5.6
-import { DiagnosticPerformanceData as TsDiagnosticPerformanceData } from '../tsServer/protocol/protocol';
+import { ResourceMap } from '../utils/resourceMap';
 
 function diagnosticsEquals(a: vscode.Diagnostic, b: vscode.Diagnostic): boolean {
 	if (a === b) {
@@ -217,13 +216,9 @@ class DiagnosticsTelemetryManager extends Disposable {
 			*/
 			this._telemetryReporter.logTelemetry('diagnostics.performance',
 				{
-					// @ts-expect-error until ts 5.6
 					syntaxDiagDuration: data.syntaxDiag,
-					// @ts-expect-error until ts 5.6
 					semanticDiagDuration: data.semanticDiag,
-					// @ts-expect-error until ts 5.6
 					suggestionDiagDuration: data.suggestionDiag,
-					// @ts-expect-error until ts 5.6
 					regionSemanticDiagDuration: data.regionSemanticDiag,
 					fileLineCount: data.fileLineCount,
 				},
