@@ -26,7 +26,6 @@ export namespace Range {
 			Math.max(0, start.line - 1), Math.max(start.offset - 1, 0),
 			Math.max(0, end.line - 1), Math.max(0, end.offset - 1));
 
-	// @ts-expect-error until ts 5.6
 	export const toFileRange = (range: vscode.Range): Proto.FileRange => ({
 		startLine: range.start.line + 1,
 		startOffset: range.start.character + 1,
@@ -38,7 +37,7 @@ export namespace Range {
 		file,
 		...toFileRange(range)
 	});
-	// @ts-expect-error until ts 5.6
+
 	export const toFileRangesRequestArgs = (file: string, ranges: vscode.Range[]): Proto.FileRangesRequestArgs => ({
 		file,
 		ranges: ranges.map(toFileRange)
