@@ -717,16 +717,13 @@ export abstract class Layout extends Disposable implements IWorkbenchLayoutServi
 		}
 
 		// Auxiliary Panel to restore
-		const auxiliaryBarVisible = this.isVisible(Parts.AUXILIARYBAR_PART);
-		if (auxiliaryBarVisible || (isNewWindow && activityBarNotDefault)) {
+		if (this.isVisible(Parts.AUXILIARYBAR_PART)) {
 			const viewContainerToRestore = this.storageService.get(AuxiliaryBarPart.activePanelSettingsKey, StorageScope.WORKSPACE, this.viewDescriptorService.getDefaultViewContainer(ViewContainerLocation.AuxiliaryBar)?.id);
 
 			if (viewContainerToRestore) {
 				this.state.initialization.views.containerToRestore.auxiliaryBar = viewContainerToRestore;
-			}
-
-			if (auxiliaryBarVisible !== !!viewContainerToRestore) {
-				this.stateModel.setRuntimeValue(LayoutStateKeys.AUXILIARYBAR_HIDDEN, !auxiliaryBarVisible);
+			} else {
+				this.stateModel.setRuntimeValue(LayoutStateKeys.AUXILIARYBAR_HIDDEN, true);
 			}
 		}
 
