@@ -24,6 +24,8 @@ import { IConfigurationService } from '../../../../../platform/configuration/com
 import { IUnchangedEditorRegionsService } from './unchangedEditorRegions.js';
 import { Schemas } from '../../../../../base/common/network.js';
 
+const PropertyHeaderHeight = 25;
+
 // From `.monaco-editor .diff-hidden-lines .center` in src/vs/editor/browser/widget/diffEditor/style.css
 export const HeightOfHiddenLinesRegionInDiffEditor = 24;
 
@@ -47,6 +49,7 @@ export type IDiffElementViewModelBase = DiffElementCellViewModelBase | DiffEleme
 export abstract class DiffElementViewModelBase extends Disposable {
 	protected _layoutInfoEmitter = this._register(new Emitter<CellDiffViewModelLayoutChangeEvent>());
 	onDidLayoutChange = this._layoutInfoEmitter.event;
+	abstract renderOutput: boolean;
 	constructor(
 		public readonly mainDocumentTextModel: INotebookTextModel,
 		public readonly editorEventDispatcher: NotebookDiffEditorEventDispatcher,
@@ -72,6 +75,7 @@ export class DiffElementPlaceholderViewModel extends DiffElementViewModelBase {
 	protected _unfoldHiddenCells = this._register(new Emitter<void>());
 	onUnfoldHiddenCells = this._unfoldHiddenCells.event;
 
+	public renderOutput: boolean = false;
 	constructor(
 		mainDocumentTextModel: INotebookTextModel,
 		editorEventDispatcher: NotebookDiffEditorEventDispatcher,
@@ -223,17 +227,17 @@ export abstract class DiffElementCellViewModelBase extends DiffElementViewModelB
 		this.original = original ? this._register(new DiffNestedCellViewModel(original, notebookService)) : undefined;
 		this.modified = modified ? this._register(new DiffNestedCellViewModel(modified, notebookService)) : undefined;
 		const editorHeight = this._estimateEditorHeight(initData.fontInfo);
-		const cellStatusHeight = 25;
+		const cellStatusHeight = PropertyHeaderHeight;
 		this._layoutInfo = {
 			width: 0,
 			editorHeight: editorHeight,
 			editorMargin: 0,
 			metadataHeight: 0,
 			cellStatusHeight,
-			metadataStatusHeight: this.ignoreMetadata ? 0 : 25,
+			metadataStatusHeight: this.ignoreMetadata ? 0 : PropertyHeaderHeight,
 			rawOutputHeight: 0,
 			outputTotalHeight: 0,
-			outputStatusHeight: this.ignoreOutputs ? 0 : 25,
+			outputStatusHeight: this.ignoreOutputs ? 0 : PropertyHeaderHeight,
 			outputMetadataHeight: 0,
 			bodyMargin: 32,
 			totalHeight: 82 + cellStatusHeight + editorHeight,
